update product news controller

+++ app/Http/Controllers/ProductController.php
@@ -4,8 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Menu;
 use App\Models\Product;
+use App\Models\ProductMenu;
 use GuzzleHttp\Handler\Proxy;
 use Illuminate\Http\Request;
+use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Pagination\Paginator;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\File;
 use Illuminate\Support\Str;
 
 class ProductController extends Controller
@@ -17,9 +22,12 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $products = Product::all();
         $menus = Menu::all();
-        return view('admin.product.index')->with('products', $products)->with('menus', $menus);
+        $products = DB::table('products')->join('menu_product', 'menu_product.product_id', '=', 'products.id')
+                        ->select('products.*', 'menu_product.priority')
+                        ->orderBy('menu_product.priority')
+                        ->paginate(2);
+        return view('admin.product.index', ['products' => $products])->with('menus', $menus);
     }
 
     /**
@@ -29,8 +37,9 @@ class ProductController extends Controller
      */
     public function create()
     {
-        //
-        return view('admin.product.create');
+        $menus = Menu::all();
+
+        return view('admin.product.create', ['menus' => $menus]);
     }
 
     /**
@@ -41,10 +50,18 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        $product_menus = ProductMenu::where('subcategory_id', $request->input('menu_id'))
+            ->where('priority', $request->input('priority'))
+            ->first();
+        // dd($request);
+        if ($product_menus != null) {
+            return redirect()->back()->withErrors(['error' => 'Thứ tự ưu tiên đã tồn tại']);
+        }
+
         $name = '';
         $temp = str_split($request->input('name'));
         foreach ($temp as $char) {
-            if($char == ' ') {
+            if ($char == ' ') {
                 $char = '-';
             }
             $name .= $char;
@@ -52,18 +69,17 @@ class ProductController extends Controller
         $imgData = [];
         if ($request->hasfile('image')) {
             foreach ($request->image as $key => $file) {
-                if($key == 0) {
+                if ($key == 0) {
                     $image_url = time() . $name . '.' . $file->extension();
                     $imgData[] = $image_url;
-                }
-                else{
+                } else {
                     $image_url = time() . $name . '(' . $key . ')' . '.' . $file->extension();
                     $imgData[] = $image_url;
                 }
             }
         }
         $product = new Product();
-        if($imgData) {
+        if ($imgData) {
             foreach ($imgData as $key => $image) {
                 $name = 'image_' . ($key + 1);
                 $product->$name = $image;
@@ -87,11 +103,18 @@ class ProductController extends Controller
 
         $success = $product->save();
 
-        if ($success) {
+        if ($success && $request->image) {
             foreach ($request->image as $key => $file) {
                 $file->move(public_path('upload/images/product'), $imgData[$key]);
             }
         }
+
+
+        $product_menu = new ProductMenu();
+        $product_menu->subcategory_id = $request->input('menu_id');
+        $product_menu->product_id = $product->id;
+        $product_menu->priority = $request->input('priority');
+        $product_menu->save();
 
         return redirect()->route('admin.product.index');
     }
@@ -113,9 +136,14 @@ class ProductController extends Controller
      * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function edit(Product $product)
+    public function edit($id)
     {
-        return view('admin.product.edit', ['product' => $product]);
+        $product_edit = Product::join('menu_product', 'menu_product.product_id', '=', 'products.id')
+            ->where('products.id', $id)
+            ->get(['products.id', 'products.name', 'products.code', 'products.title', 'products.description', 'products.content', 'products.specifications', 'products.sell_price', 'products.sale_price', 'products.size', 'products.sold_out', 'products.guarantee', 'products.status', 'products.image_1', 'products.image_2', 'products.image_3', 'products.is_contact_product', 'products.is_sale_in_month', 'products.is_hot_product', 'products.created_at', 'menu_product.priority',])->first();
+        $menus = Menu::all();
+
+        return view('admin.product.edit', ['product' => $product_edit], ['menus' => $menus]);
     }
 
     /**
@@ -127,10 +155,16 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+        $product_menus = ProductMenu::where('subcategory_id', $request->input('subcategory_id'))
+            ->where('priority', $request->input('priority'))
+            ->first();
+        if ($product_menus && $product_menus->priority != $request->input('priority')) {
+            return redirect()->back()->withErrors(['error' => 'Thứ tự ưu tiên đã tồn tại']);
+        }
         $name = '';
         $temp = str_split($request->input('name'));
         foreach ($temp as $char) {
-            if($char == ' ') {
+            if ($char == ' ') {
                 $char = '-';
             }
             $name .= $char;
@@ -138,37 +172,38 @@ class ProductController extends Controller
         $imgData = [];
         if ($request->hasfile('image')) {
             foreach ($request->image as $key => $file) {
-                if($key == 0) {
+                if ($key == 0) {
+                    if($product->image_1){
+                        if (File::exists(public_path('upload/images/product/') . $product->image_1)) {
+                            unlink(public_path('upload/images/product/'). $product->image_1);
+                        }
+                    }
                     $image_url = time() . $name . '.' . $file->extension();
+                    $file->move(public_path('upload/images/product'), $image_url);
                     $imgData[] = $image_url;
-                }
-                else{
+                } else {
+                    $name_temp = 'image_' . ($key + 1);
+                    if($product->$name_temp){
+                        if (File::exists(public_path('upload/images/product/') . $product->$name_temp)) {
+                            unlink(public_path('upload/images/product/'). $product->$name_temp);
+                        }
+                    }
                     $image_url = time() . $name . '(' . $key . ')' . '.' . $file->extension();
+                    $file->move(public_path('upload/images/product'),  $image_url);
                     $imgData[] = $image_url;
                 }
-            }
-        }else {
-            if($product->image_1) {
-                unlink(public_path('upload/images/product/'). $product->image_1);
-            }
-            if($product->image_2) {
-                unlink(public_path('upload/images/product/'). $product->image_2);
-            }
-            if($product->image_3) {
-                unlink(public_path('upload/images/product/'). $product->image_3);
             }
         }
 
-
         $product_update = Product::find($product->id);
-        $product->image_1 ? unlink(public_path('upload/images/product/'). $product->image_1) : null;
-        $product->image_2 ? unlink(public_path('upload/images/product/'). $product->image_2) : null;
-        $product->image_3 ? unlink(public_path('upload/images/product/'). $product->image_3) : null;
-
-        if($imgData) {
+        if ($imgData) {
             foreach ($imgData as $key => $image) {
                 $name = 'image_' . ($key + 1);
-                $product_update->$name = $image;
+                if($image){
+                    $product_update->$name = $image;
+                }else{
+                    $product_update->$name = $product_update->$name;
+                }
             }
         }
         $product_update->name = $request->input('name');
@@ -187,11 +222,6 @@ class ProductController extends Controller
         $product_update->is_sale_in_month = $request->input('is_sale_in_month');
         $product_update->is_hot_product = $request->input('is_hot_product');
         $success = $product_update->update();
-        if ($success) {
-            foreach ($request->image as $key => $file) {
-                $file->move(public_path('upload/images/product'), $imgData[$key]);
-            }
-        }
         return redirect()->route('admin.product.index');
     }
 
@@ -201,22 +231,29 @@ class ProductController extends Controller
      * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Product $product)
+    public function destroy($id)
     {
+        $product = Product::find($id);
         if ($product->image_1) {
             $image_url1 = public_path('upload/images/product/') . $product->image_1;
-            unlink($image_url1);
+            if (File::exists($image_url1)) {
+                unlink($image_url1);
+            }
         }
         if ($product->image_2) {
             $image_url2 = public_path('upload/images/product/') . $product->image_2;
-            unlink($image_url2);
+            if (File::exists($image_url2)) {
+                unlink($image_url2);
+            }
         }
         if ($product->image_2) {
             $image_url3 = public_path('upload/images/product/') . $product->image_3;
-            unlink($image_url3);
+            if (File::exists($image_url3)) {
+                unlink($image_url3);
+            }
         }
 
-        Product::where('id', $product->id)->delete();
+        $product->delete();
         return redirect()->route('admin.product.index');
     }
 }
